+++ src/Screens/GamePlayPanels/MenuBar.cs
@@ -9,6 +9,7 @@
 
 using System;
 using System.Drawing;
+using System.Linq;
 using CivOne.Enums;
 using CivOne.Events;
 using CivOne.Graphics;
@@ -17,34 +18,79 @@ namespace CivOne.Screens.GamePlayPanels
 {
 	internal class MenuBar : BaseScreen
 	{
-		private static readonly int[] MenuXPositions =
-		[
-			8,
-			64,
-			128,
-			192,
-			240
-		];
+		private static bool DebugMenuEnabled => Settings.DebugMenu || RuntimeHandler.Runtime?.Settings.Get<bool>("debug") == true;
 		
-		public event EventHandler GameSelected;
-		public event EventHandler OrdersSelected;
-		public event EventHandler AdvisorsSelected;
-		public event EventHandler WorldSelected;
-		public event EventHandler CivilopediaSelected;
+		public event EventHandler? GameSelected;
+		public event EventHandler? OrdersSelected;
+		public event EventHandler? AdvisorsSelected;
+		public event EventHandler? WorldSelected;
+		public event EventHandler? CivilopediaSelected;
+		public event EventHandler? TerrainSelected;
 
 		public bool MenuDrag { get; private set; }
 		
 		private readonly Rectangle[] _rectMenus;
 		private readonly MenuBarHotkeyDelegate _menuBarHotkeyDelegate;
 		
-		private MenuBarTitle[] _menuTitles;
+		private MenuBarTitle[] _menuTitles = [];
+		private int[] _menuXPositions = [];
 		private bool _update;
 		private int _mouseX, _mouseY;
 
+		private void BuildMenuLayout()
+		{
+			if (_menuTitles.Length == 0)
+			{
+				_menuXPositions = [];
+				return;
+			}
+
+			int[] titleWidths = new int[_menuTitles.Length];
+			for (int i = 0; i < _menuTitles.Length; i++)
+			{
+				titleWidths[i] = Resources.GetTextSize(0, _menuTitles[i].VisibleText).Width;
+			}
+
+			int leftPadding = 8;
+			int rightPadding = 4;
+			int availableWidth = Math.Max(0, Width - leftPadding - rightPadding);
+			int totalTitleWidth = titleWidths.Sum();
+			int menuCount = _menuTitles.Length;
+			int gapCount = Math.Max(0, menuCount - 1);
+			int gap = (gapCount > 0)
+				? Math.Max(1, (availableWidth - totalTitleWidth) / gapCount)
+				: 0;
+
+			_menuXPositions = new int[menuCount];
+			int x = leftPadding;
+			for (int i = 0; i < menuCount; i++)
+			{
+				_menuXPositions[i] = x;
+				x += titleWidths[i] + gap;
+			}
+
+			for (int i = 0; i < _rectMenus.Length; i++)
+			{
+				if (i < menuCount)
+				{
+					int left = (i == 0) ? 0 : Math.Max(0, _menuXPositions[i] - 2);
+					int right = (i == (menuCount - 1))
+						? (Width - 1)
+						: Math.Min(Width - 1, _menuXPositions[i + 1] - 1);
+					int rectWidth = Math.Max(1, right - left + 1);
+					_rectMenus[i] = new Rectangle(left, 0, rectWidth, Height);
+				}
+				else
+				{
+					_rectMenus[i] = Rectangle.Empty;
+				}
+			}
+		}
+
 		private void RebuildMenuTitles()
 		{
-			_menuTitles =
-			[
+			var titles = new System.Collections.Generic.List<MenuBarTitle>
+			{
 				// it is necessary to provide the translation key as well as the translated text
 				// because the translated text may not contain the marker character '~' to indicate the hotkey position
 				// Futhermore the civtranslate cli will parse Translate(" to extract the translation keys.
@@ -53,7 +99,15 @@ namespace CivOne.Screens.GamePlayPanels
 				_menuBarHotkeyDelegate.Create(translationKey: "ADVISORS", translatedText: Translate("ADVISORS")),
 				_menuBarHotkeyDelegate.Create(translationKey: "WORLD", translatedText: Translate("WORLD")),
 				_menuBarHotkeyDelegate.Create(translationKey: "CIVILOPEDIA", translatedText: Translate("CIVILOPEDIA"))
-			];
+			};
+
+			if (DebugMenuEnabled)
+			{
+				titles.Add(_menuBarHotkeyDelegate.Create(translationKey: "TERRAIN", translatedText: Translate("TERRAIN")));
+			}
+
+			_menuTitles = [.. titles];
+			BuildMenuLayout();
 		}
 
 		private void EnsureMenuTitles()
@@ -100,6 +154,9 @@ namespace CivOne.Screens.GamePlayPanels
 				case 4:
 					CivilopediaSelected?.Invoke(this, EventArgs.Empty);
 					break;
+				case 5:
+					TerrainSelected?.Invoke(this, EventArgs.Empty);
+					break;
 			}
 		}
 		
@@ -109,9 +166,9 @@ namespace CivOne.Screens.GamePlayPanels
 			{
 				RebuildMenuTitles();
 				this.Clear(5);
-				for (int i = 0; i < MenuXPositions.Length; i++)
+				for (int i = 0; i < _menuTitles.Length; i++)
 				{
-					this.DrawText(_menuTitles[i].VisibleText, MenuXPositions[i], 1, TextSettings.DifferentCharacter(15, 7, _menuTitles[i].HighlightedCharacterIndex));
+					this.DrawText(_menuTitles[i].VisibleText, _menuXPositions[i], 1, TextSettings.DifferentCharacter(15, 7, _menuTitles[i].HighlightedCharacterIndex));
 				}
 
 				_update = false;
@@ -134,8 +191,9 @@ namespace CivOne.Screens.GamePlayPanels
 		{
 			_mouseX = args.X;
 			_mouseY = args.Y;
+			EnsureMenuTitles();
 
-			for (int i = 0; i < _rectMenus.Length; i++)
+			for (int i = 0; i < _menuTitles.Length; i++)
 			{
 				if (_rectMenus[i].Contains(args.Location))
 				{
@@ -167,12 +225,13 @@ namespace CivOne.Screens.GamePlayPanels
 
 			DefaultTextSettings = TextSettings.DifferentCharacter(15, 7, 0);
 			
-			_rectMenus = new Rectangle[5];
-			_rectMenus[0] = new Rectangle(0, 0, 56, 8);
-			_rectMenus[1] = new Rectangle(56, 0, 64, 8);
-			_rectMenus[2] = new Rectangle(120, 0, 64, 8);
-			_rectMenus[3] = new Rectangle(184, 0, 48, 8);
-			_rectMenus[4] = new Rectangle(232, 0, 88, 8);
+			_rectMenus = new Rectangle[6];
+			_rectMenus[0] = new Rectangle(0, 0, 50, 8);
+			_rectMenus[1] = new Rectangle(50, 0, 54, 8);
+			_rectMenus[2] = new Rectangle(104, 0, 54, 8);
+			_rectMenus[3] = new Rectangle(158, 0, 54, 8);
+			_rectMenus[4] = new Rectangle(212, 0, 54, 8);
+			_rectMenus[5] = new Rectangle(266, 0, 54, 8);
 		}
 	}
 }